evm: move EVMResult, ExecResult into evm types
