+++ packages/evm/src/index.ts
@@ -1,10 +1,10 @@
 import { EOF } from './eof.js'
-import { EVM, EVMResult, ExecResult } from './evm.js'
+import { EVM } from './evm.js'
 import { ERROR as EVMErrorMessage, EvmError } from './exceptions.js'
 import { InterpreterStep } from './interpreter.js'
 import { Message } from './message.js'
 import { PrecompileInput, getActivePrecompiles } from './precompiles/index.js'
-import { EVMInterface, Log } from './types.js'
+import { EVMInterface, EVMResult, ExecResult, Log } from './types.js'
 export {
   EOF,
   EVM,
