+++ packages/evm/src/precompiles/04-identity.ts
@@ -2,7 +2,7 @@ import { short } from '@ethereumjs/util'
 
 import { OOGResult } from '../evm.js'
 
-import type { ExecResult } from '../evm.js'
+import type { ExecResult } from '../types.js'
 import type { PrecompileInput } from './types.js'
 
 export function precompile04(opts: PrecompileInput): ExecResult {
