+++ packages/evm/src/precompiles/06-ecadd.ts
@@ -3,7 +3,7 @@ import { ec_add } from 'rustbn-wasm'
 
 import { OOGResult } from '../evm.js'
 
-import type { ExecResult } from '../evm.js'
+import type { ExecResult } from '../types.js'
 import type { PrecompileInput } from './types.js'
 
 export function precompile06(opts: PrecompileInput): ExecResult {
