+++ packages/evm/src/precompiles/0a-kzg-point-evaluation.ts
@@ -11,7 +11,7 @@ import {
 import { EvmErrorResult, OOGResult } from '../evm.js'
 import { ERROR, EvmError } from '../exceptions.js'
 
-import type { ExecResult } from '../evm.js'
+import type { ExecResult } from '../types.js'
 import type { PrecompileInput } from './types.js'
 
 export const BLS_MODULUS = BigInt(
